Allow customising options passed to keychain
This also changes the default to be just `--quiet`.  All the other
options used by the plugin before are in fact a matter of preference and
IMHO shouldn't be default, e.g. `--confhost` produced a warning if
config file isn't used, `--timeout` can be annoying, and GPG isn't used
on every system.